+++ plugin.video.watchmojo/addon.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<addon id="plugin.video.watchmojo" name="Watchmojo" provider-name="willforde" version="3.0.1">
+<addon id="plugin.video.watchmojo" name="Watchmojo" provider-name="willforde" version="3.0.2">
     <requires>
         <import addon="xbmc.python" version="2.25.0"/>
         <import addon="script.module.codequick" version="0.9.2"/>
@@ -24,6 +24,6 @@
             <screenshot>resources/screenshots/screenshot001.jpg</screenshot>
             <screenshot>resources/screenshots/screenshot002.jpg</screenshot>
         </assets>
-        <news>[fix] Fixed playback problem with Kodi on Xbox.\n[fix] Fixed SSL verification failure, preventing add-on from working.\n[upd] Add WatchMojo UK channel.</news>
+        <news>[fix] Fix some scraping issues that added some useless entries.</news>
     </extension>
 </addon>
